fix PR feedback and making calculatedSpeed a Double

+++ service/src/main/java/eu/europa/ec/fisheries/uvms/movement/service/bean/MovementCreateBean.java
@@ -101,7 +101,7 @@ public class MovementCreateBean {
             movement.setMovementConnect(movementConnect);
             
             if (previousVms != null && !MovementSourceType.AIS.value().equals(movement.getSource().value()) ) {
-                movement.setCalculatedSpeed((float)CalculationUtil.getPositionCalculations(previousVms, movement).getAvgSpeed());
+                movement.setCalculatedSpeed(CalculationUtil.getPositionCalculations(previousVms, movement).getAvgSpeed());
             }
 
             Movement createdMovement = movementService.createAndProcessMovement(movement);
